data_management: reorg object pointer
